# packages/api-node/openapi.yaml
openapi: 3.0.3
info:
  title: User Profile App - API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
...
  /api/auth/register:
    post:
      summary: Register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                displayName: { type: string }
      responses:
        '201': { description: Created }
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '200': { description: OK }
...
  /api/users:
    get:
      summary: Search users
      tags: [Users]
      parameters:
        - in: query
          name: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...